create database hw
