+++ .gitignore
@@ -2,3 +2,5 @@
 /homework_async/TP_AsyncCall/TP_AsyncCall-ejb/target/
 /homework_async/TP_AsyncCall/TP_AsyncCall-web/target/
 /homework_database/tp_homework_database-bkp/target/
+/homwork_database/TP_database/target/
+/homework_database/TP_db/target/
