--- homework_async/TP_AsyncCall/.gitignore
@@ -1,2 +1,0 @@
-nb-configuration.xml
-target/
