+++ homework_database/TP_db/src/main/java/hu/codingmentor/tp_db/ColorType.java
@@ -1,0 +1,5 @@
+package hu.codingmentor.tp_db;
+
+public enum ColorType {
+    GREEN, RED, BLACK, WHITE, YELLOW;
+}
