+++ homework_database/TP_db/src/main/java/hu/codingmentor/tp_db/Gender.java
@@ -1,0 +1,5 @@
+package hu.codingmentor.tp_db;
+
+public enum Gender {
+    FEMALE, MALE;
+}
