+++ homework_database/TP_db/src/main/java/hu/codingmentor/tp_db/LengthOfItems.java
@@ -1,0 +1,6 @@
+package hu.codingmentor.tp_db;
+
+public enum LengthOfItems {
+SHORT,LONG;
+}
+
