--- homework_database/tp_homework_database-bkp/src/main/java/hu/codingmentor/tp/homework/database/Gender.java
@@ -1,6 +1,0 @@
-
-package hu.codingmentor.tp.homework.database;
-
-public enum Gender {
-    FEMALE, MALE;
-}
