--- homework_database/tp_homework_database-bkp/src/main/java/hu/codingmentor/tp/homework/database/LengthOfItems.java
@@ -1,8 +1,0 @@
-
-
-package hu.codingmentor.tp.homework.database;
-
-
-public enum LengthOfItems {
-SHORT,LONG;
-}
