--- homework_database/tp_homework_database/src/main/java/hu/codingmentor/tp/homework/database/Article.java
@@ -1,31 +1,0 @@
-
-package hu.codingmentor.tp.homework.database;
-
-import java.io.Serializable;
-import javax.persistence.DiscriminatorValue;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.NamedQuery;
-
-@Entity
-@DiscriminatorValue("B")
-@NamedQuery(name = "QueryArticle" , query = "SELECT b FROM Article b WHERE b.lengthOfItem=hu.codingmentor.tp_homework_database.LengthOfItems.SHORT")
-public class Article extends Item implements Serializable {
-
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    public Article() {
-        //default
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-}
