--- homework_database/tp_homework_database/src/main/java/hu/codingmentor/tp/homework/database/ColorType.java
@@ -1,6 +1,0 @@
-
-package hu.codingmentor.tp.homework.database;
-
-public enum ColorType {
-    GREEN, RED, BLACK, WHITE, YELLOW;
-}
